address review

+++ oneflow/core/common/util.cpp
@@ -129,9 +129,9 @@ bool StringToInteger(const std::string& str, int64_t* value) {
   }
 }
 
-bool StringToFloat(const std::string& str, float* value) {
+bool StringToFloat(const std::string& str, double* value) {
   char* end = nullptr;
-  float v = std::strtof(str.data(), &end);
+  double v = std::strtof(str.data(), &end);
   if (end == str.data()) {
     return false;
   } else {
@@ -162,10 +162,10 @@ int64_t ParseIntegerFromEnv(const std::string& env_var, int64_t default_value) {
   }
 }
 
-float ParseFloatFromEnv(const std::string& env_var, float default_value) {
+double ParseFloatFromEnv(const std::string& env_var, double default_value) {
   const char* env_p = std::getenv(env_var.c_str());
   if (env_p == nullptr) { return default_value; }
-  float value = default_value;
+  double value = default_value;
   StringToFloat(env_p, &value);
   return value;
 }
